+++ changelog.md
@@ -3,6 +3,8 @@
 * Allow spawners to be empowered with echo shards
 
 ## Upcoming
+* Added Vigour enchantment
+  * A new chestplate enchantment that reduces stamina consumption
 * OverTime healing now decays based off exhaustion instead of every tick
 * Supplementaries 
   * Flax is now hard to break and grows slower. Also no longer drops more than 1 seed